feat(xhr): Support responseType=blob + string
In order to support the `whatwg-fetch` polyfill which sets
`responseType='blob'` for every request, create a new Blob
when the responseType is `blob` and the body is a string.

Added integration tests for `whatwg-fetch`

Fixed proxy tests that are failing because the remote services response
have changed. `reqres.in` and `httpbin.com` have changed their behavior.
Fix those for a working test suite again.

Typescript definitions needed some updates in order to pass. Upgraded
Typescript to 4.0.